classroom_management_streamlit/requirements.txt
streamlit==1.44.1
pymongo==4.11.3
pandas==2.2.3
python-dotenv==1.0.1
